feat: refactor utils ladder dir
