+++ src/renderer/components/_molecules/rung/ladder-utils/nodes.ts
@@ -2,65 +2,67 @@ import type { FlowState } from '@root/renderer/store/slices'
 import type { Edge, Node } from '@xyflow/react'
 import { getNodesBounds } from '@xyflow/react'
 
-import { customNodesStyles, nodesBuilder } from '../../_atoms/react-flow/custom-nodes'
-import { CustomHandleProps } from '../../_atoms/react-flow/custom-nodes/handle'
+import { customNodesStyles, nodesBuilder } from '../../../_atoms/react-flow/custom-nodes'
+import type { CustomHandleProps } from '../../../_atoms/react-flow/custom-nodes/handle'
+import { connectNodes, disconnectNodes } from './edges'
 
-/**
- * EDGES
- */
-const buildEdge = (sourceNodeId: string, targetNodeId: string): Edge => {
-  return {
-    id: `e_${sourceNodeId}_${targetNodeId}`,
-    source: sourceNodeId,
-    target: targetNodeId,
-    type: 'step',
+export const getNodeStyle = ({ node, nodeType }: { node?: Node; nodeType?: string }) => {
+  return customNodesStyles[node?.type ?? nodeType ?? 'mockNode']
+}
+
+export const nodeBuild = ({
+  nodeType,
+  id,
+  posX,
+  posY,
+  handleX,
+  handleY,
+}: {
+  nodeType: string
+  id: string
+  posX: number
+  posY: number
+  handleX: number
+  handleY: number
+}) => {
+  switch (nodeType) {
+    case 'block':
+      return nodesBuilder.block({
+        id,
+        posX,
+        posY,
+        handleX,
+        handleY,
+      })
+    case 'coil':
+      return nodesBuilder.coil({
+        id,
+        posX,
+        posY,
+        handleX,
+        handleY,
+      })
+    case 'contact':
+      return nodesBuilder.contact({
+        id,
+        posX,
+        posY,
+        handleX,
+        handleY,
+      })
+    default:
+      // mock block
+      return nodesBuilder.mockNode({
+        id,
+        label: id,
+        posX,
+        posY,
+        handleX,
+        handleY,
+      })
   }
 }
 
-const removeEdge = (edges: Edge[], edgeId: string): Edge[] => {
-  return edges.filter((edge) => edge.id !== edgeId)
-}
-
-export const connectNodes = (rung: FlowState, sourceNodeId: string, targetNodeId: string): Edge[] => {
-  // Find the source edge
-  const sourceEdge = rung.edges.find((edge) => edge.source === sourceNodeId)
-
-  // If the source edge is found, update the target
-  if (sourceEdge) {
-    // Remove the source edge
-    const edges = rung.edges.filter((edge) => edge.id !== sourceEdge.id)
-    // Update the target of the source edge
-    edges.push(buildEdge(sourceNodeId, targetNodeId))
-    // Update the target of the target edge
-    edges.push(buildEdge(targetNodeId, sourceEdge.target))
-    return edges
-  }
-
-  const edges = rung.edges
-  return [...edges, buildEdge(sourceNodeId, targetNodeId)]
-}
-
-export const disconnectNodes = (rung: FlowState, sourceNodeId: string, targetNodeId: string): Edge[] => {
-  // sourceEdge -> edge that connect the source node
-  // targetEdge -> edge that connect the target node
-  const sourceEdge = rung.edges.find((edge) => edge.target === sourceNodeId)
-  const targetEdge = rung.edges.find((edge) => edge.target === targetNodeId)
-
-  if (!sourceEdge) return rung.edges
-
-  let newEdges = removeEdge(rung.edges, sourceEdge.id)
-
-  if (targetEdge) {
-    newEdges = removeEdge(newEdges, targetEdge.id)
-    newEdges.push(buildEdge(sourceEdge.source, targetEdge.target))
-  }
-
-  return newEdges
-}
-
-/**
- * NODES
- */
 export const changePowerRailBounds = ({
   rung,
   nodes,
@@ -102,64 +104,6 @@ export const changePowerRailBounds = ({
   }
 
   return newRail
-}
-
-export const getNodeStyle = ({ node, nodeType }: { node?: Node; nodeType?: string }) => {
-  return customNodesStyles[node?.type ?? nodeType ?? 'mockNode']
-}
-
-export const nodeBuild = ({
-  nodeType,
-  id,
-  posX,
-  posY,
-  handleX,
-  handleY,
-}: {
-  nodeType: string
-  id: string
-  posX: number
-  posY: number
-  handleX: number
-  handleY: number
-}) => {
-  switch (nodeType) {
-    case 'block':
-      return nodesBuilder.block({
-        id,
-        posX,
-        posY,
-        handleX,
-        handleY,
-        variant: 'TON',
-      })
-    case 'coil':
-      return nodesBuilder.coil({
-        id,
-        posX,
-        posY,
-        handleX,
-        handleY,
-      })
-    case 'contact':
-      return nodesBuilder.contact({
-        id,
-        posX,
-        posY,
-        handleX,
-        handleY,
-      })
-    default:
-      // mock block
-      return nodesBuilder.mockNode({
-        id,
-        label: id,
-        posX,
-        posY,
-        handleX,
-        handleY,
-      })
-  }
 }
 
 export const addNewNode = ({
